# web/openapi.json
{
  "components": {
    "schemas": {
...
      "ApprovalItemRef": {
        "description": "待审记录的**只读**引用（完整契约在 `app/schemas/approvals.py`）。\n\n这里不直接 import 那个模块：`schemas` 之间横向依赖会让\"改一处、炸两处\"，\n而面板需要的只是\"当前这条待审的 id / 分数 / 轮次\"。",
        "properties": {
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "requested_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested At"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "score_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Total"
          },
          "script_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Script Id"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "task_id",
          "status"
        ],
        "title": "ApprovalItemRef",
        "type": "object"
      },
...
      "DiffSummaryModel": {
        "description": "改动计数（面板顶部的\"改 3 句 / 新增 1 句 / 删 2 句\"）。",
        "properties": {
          "added": {
            "default": 0,
            "title": "Added",
            "type": "integer"
          },
          "changed": {
            "default": 0,
            "title": "Changed",
            "type": "integer"
          },
          "removed": {
            "default": 0,
            "title": "Removed",
            "type": "integer"
          },
          "total": {
            "default": 0,
            "title": "Total",
            "type": "integer"
          },
          "unchanged": {
            "default": 0,
            "title": "Unchanged",
            "type": "integer"
          }
        },
        "title": "DiffSummaryModel",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ScriptBody": {
        "description": "一版稿子的正文（不含逐句表 —— 那个单独给，面板要分段渲染）。",
        "properties": {
          "body_md": {
            "title": "Body Md",
            "type": "string"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "cta": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cta"
          },
          "editor_notes": {
            "items": {},
            "title": "Editor Notes",
            "type": "array"
          },
          "est_duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Est Duration Ms"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "hook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hook"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "outline": {
            "additionalProperties": true,
            "title": "Outline",
            "type": "object"
          },
          "prompt_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Version"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "score_llm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Llm"
          },
          "score_rule": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Rule"
          },
          "score_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Total"
          },
          "speaker_ratio": {
            "additionalProperties": true,
            "title": "Speaker Ratio",
            "type": "object"
          },
          "target_chars": {
            "default": 700,
            "title": "Target Chars",
            "type": "integer"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "word_count": {
            "default": 0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version",
          "is_active",
          "body_md",
          "editor_notes",
          "outline",
          "speaker_ratio"
        ],
        "title": "ScriptBody",
        "type": "object"
      },
      "ScriptDetail": {
        "description": "稿件面板的整页数据（一次请求拿全，避免面板\"拼\"出不一致的视图）。\n\n:param reviews: **全部**轮次（按轮次正序）。面板要能并排显示\n    \"第 1 轮 6.3 分 / 第 2 轮 7.1 分\"，否则\"改稿有没有用\"永远说不清。\n:param approval: 当前待审记录；不在闸里 ⇒ ``None``（面板据此决定是否显示决断条）。",
        "properties": {
          "approval": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApprovalItemRef"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/ScriptReview"
            },
            "title": "Reviews",
            "type": "array"
          },
          "revision_round": {
            "title": "Revision Round",
            "type": "integer"
          },
          "script": {
            "$ref": "#/components/schemas/ScriptBody"
          },
          "sentences": {
            "items": {
              "$ref": "#/components/schemas/ScriptSentence"
            },
            "title": "Sentences",
            "type": "array"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "task_status": {
            "title": "Task Status",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "task_status",
          "revision_round",
          "script",
          "sentences",
          "reviews"
        ],
        "title": "ScriptDetail",
        "type": "object"
      },
      "ScriptDiff": {
        "description": "两个版本之间的逐句对照。",
        "properties": {
          "changes": {
            "items": {
              "$ref": "#/components/schemas/SentenceDiffRow"
            },
            "title": "Changes",
            "type": "array"
          },
          "from_version": {
            "title": "From Version",
            "type": "integer"
          },
          "summary": {
            "$ref": "#/components/schemas/DiffSummaryModel"
          },
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "to_version": {
            "title": "To Version",
            "type": "integer"
          }
        },
        "required": [
          "task_id",
          "from_version",
          "to_version",
          "changes",
          "summary"
        ],
        "title": "ScriptDiff",
        "type": "object"
      },
      "ScriptReview": {
        "description": "一轮审稿结论（一次 ``review_scores`` 行）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "decision": {
            "title": "Decision",
            "type": "string"
          },
          "grade": {
            "title": "Grade",
            "type": "string"
          },
          "issues": {
            "items": {},
            "title": "Issues",
            "type": "array"
          },
          "llm_detail": {
            "additionalProperties": true,
            "title": "Llm Detail",
            "type": "object"
          },
          "llm_model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llm Model"
          },
          "llm_total": {
            "title": "Llm Total",
            "type": "number"
          },
          "round_no": {
            "title": "Round No",
            "type": "integer"
          },
          "rule_detail": {
            "additionalProperties": true,
            "title": "Rule Detail",
            "type": "object"
          },
          "rule_total": {
            "title": "Rule Total",
            "type": "number"
          },
          "total": {
            "title": "Total",
            "type": "number"
          }
        },
        "required": [
          "round_no",
          "rule_total",
          "rule_detail",
          "llm_total",
          "llm_detail",
          "total",
          "grade",
          "decision",
          "issues"
        ],
        "title": "ScriptReview",
        "type": "object"
      },
      "ScriptSentence": {
        "description": "一句待配音文本（``script_sentences`` 的展示字段）。",
        "properties": {
          "emotion": {
            "default": "neutral",
            "title": "Emotion",
            "type": "string"
          },
          "pause_after_ms": {
            "default": 200,
            "title": "Pause After Ms",
            "type": "integer"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "speaker": {
            "title": "Speaker",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "text",
          "speaker"
        ],
        "title": "ScriptSentence",
        "type": "object"
      },
      "ScriptVersion": {
        "description": "版本列表的一项（\"v1 → v2\"下拉用）。",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "grade": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grade"
          },
          "is_active": {
            "title": "Is Active",
            "type": "boolean"
          },
          "revision_round": {
            "default": 0,
            "title": "Revision Round",
            "type": "integer"
          },
          "score_total": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Total"
          },
          "sentence_count": {
            "default": 0,
            "title": "Sentence Count",
            "type": "integer"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "word_count": {
            "default": 0,
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "version",
          "is_active"
        ],
        "title": "ScriptVersion",
        "type": "object"
      },
      "ScriptVersionList": {
        "properties": {
          "task_id": {
            "title": "Task Id",
            "type": "string"
          },
          "versions": {
            "items": {
              "$ref": "#/components/schemas/ScriptVersion"
            },
            "title": "Versions",
            "type": "array"
          }
        },
        "required": [
          "task_id",
          "versions"
        ],
        "title": "ScriptVersionList",
        "type": "object"
      },
...
      "SentenceDiffRow": {
        "description": "两版之间的一句对照。",
        "properties": {
          "new_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Seq"
          },
          "new_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "New Text"
          },
          "old_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Seq"
          },
          "old_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Old Text"
          },
          "op": {
            "title": "Op",
            "type": "string"
          }
        },
        "required": [
          "op"
        ],
        "title": "SentenceDiffRow",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "AI 全自动营销号制片台",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/scripts/{task_id}": {
      "get": {
        "description": "稿件全文 + 逐句表 + 全部轮次评分 + 当前待审记录。\n\n任务不存在 ⇒ 404（``TASK_NOT_FOUND``）；任务在但还没有稿件 ⇒ 404\n（``REVIEW_SCRIPT_MISSING``）—— 两种\"没有\"分开报，面板才能给出不同的提示。",
        "operationId": "get_script_api_v1_scripts__task_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Script",
        "tags": [
          "scripts"
        ]
      }
    },
    "/api/v1/scripts/{task_id}/diff": {
      "get": {
        "description": "两版逐句对照（改稿到底改了什么）。\n\n两个版本号都必须**显式**给：默认成\"上一版 vs 当前版\"看起来很贴心，但改稿\n可能一次跳两版（人工退回 + 自动重跑），默认值会让人对着错误的对照下结论。",
        "operationId": "diff_script_versions_api_v1_scripts__task_id__diff_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          },
          {
            "description": "旧版本号",
            "in": "query",
            "name": "from_version",
            "required": true,
            "schema": {
              "description": "旧版本号",
              "minimum": 1,
              "title": "From Version",
              "type": "integer"
            }
          },
          {
            "description": "新版本号",
            "in": "query",
            "name": "to_version",
            "required": true,
            "schema": {
              "description": "新版本号",
              "minimum": 1,
              "title": "To Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptDiff"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Diff Script Versions",
        "tags": [
          "scripts"
        ]
      }
    },
    "/api/v1/scripts/{task_id}/versions": {
      "get": {
        "description": "全部版本（\"v1 → v2\"下拉用；``is_active`` 标出当前生效的那一版）。",
        "operationId": "list_script_versions_api_v1_scripts__task_id__versions_get",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScriptVersionList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Script Versions",
        "tags": [
          "scripts"
        ]
      }
    },
...
  }
}